Refactor the reconciler module
