+++ src/controller_examples/rabbitmq_controller/proof/terminate.rs
@@ -31,7 +31,7 @@ use crate::rabbitmq_controller::{
     proof::{common::*, safety::*},
     spec::{rabbitmqcluster::*, reconciler::*},
 };
-use crate::reconciler::spec::*;
+use crate::reconciler::spec::reconciler::*;
 use crate::temporal_logic::{defs::*, rules::*};
 use vstd::prelude::*;
