+++ src/controller_examples/simple_controller/exec/reconciler.rs
@@ -5,7 +5,7 @@ use crate::kubernetes_api_objects::{
     api_method::*, common::*, config_map::*, object_meta::*, resource::ResourceView,
 };
 use crate::pervasive_ext::string_map::StringMap;
-use crate::reconciler::exec::*;
+use crate::reconciler::exec::reconciler::*;
 use crate::simple_controller::common::*;
 use crate::simple_controller::spec::custom_resource::*;
 use crate::simple_controller::spec::reconciler as simple_spec;
