+++ src/controller_examples/simple_controller/spec/reconciler.rs
@@ -6,7 +6,7 @@ use crate::kubernetes_api_objects::{
     resource::*,
 };
 use crate::kubernetes_cluster::spec::message::*;
-use crate::reconciler::spec::*;
+use crate::reconciler::spec::reconciler::*;
 use crate::simple_controller::common::*;
 use crate::simple_controller::spec::custom_resource::*;
 use crate::state_machine::{action::*, state_machine::*};
