+++ src/controller_examples/zookeeper_controller/exec/reconciler.rs
@@ -8,7 +8,7 @@ use crate::kubernetes_api_objects::{
 };
 use crate::pervasive_ext::string_map::StringMap;
 use crate::pervasive_ext::string_view::*;
-use crate::reconciler::exec::*;
+use crate::reconciler::exec::{io::*, reconciler::*};
 use crate::zookeeper_controller::common::*;
 use crate::zookeeper_controller::exec::zookeepercluster::*;
 use crate::zookeeper_controller::spec::reconciler as zk_spec;
