+++ src/controller_examples/zookeeper_controller/spec/reconciler.rs
@@ -8,7 +8,7 @@ use crate::kubernetes_api_objects::{
 };
 use crate::kubernetes_cluster::spec::message::*;
 use crate::pervasive_ext::string_view::*;
-use crate::reconciler::spec::*;
+use crate::reconciler::spec::{io::*, reconciler::*};
 use crate::state_machine::{action::*, state_machine::*};
 use crate::temporal_logic::defs::*;
 use crate::zookeeper_controller::common::*;
