+++ src/kubernetes_cluster/proof/cluster.rs
@@ -16,7 +16,7 @@ use crate::kubernetes_cluster::spec::{
     network,
     network::{network, NetworkState},
 };
-use crate::reconciler::spec::Reconciler;
+use crate::reconciler::spec::reconciler::Reconciler;
 use crate::state_machine::action::*;
 use crate::state_machine::state_machine::*;
 use crate::temporal_logic::defs::*;
