+++ src/kubernetes_cluster/proof/controller_runtime_eventual_safety.rs
@@ -17,7 +17,7 @@ use crate::kubernetes_cluster::{
         message::*,
     },
 };
-use crate::reconciler::spec::Reconciler;
+use crate::reconciler::spec::reconciler::Reconciler;
 use crate::temporal_logic::defs::*;
 use crate::temporal_logic::rules::*;
 use vstd::prelude::*;
