+++ src/kubernetes_cluster/spec/cluster.rs
@@ -18,7 +18,7 @@ use crate::kubernetes_cluster::spec::{
     message::*,
     network::{network, NetworkState},
 };
-use crate::reconciler::spec::Reconciler;
+use crate::reconciler::spec::reconciler::Reconciler;
 use crate::state_machine::{action::*, state_machine::*};
 use crate::temporal_logic::defs::*;
 use vstd::{multiset::*, prelude::*};
