+++ src/reconciler/exec/external.rs
@@ -1,5 +1,2 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
-
-pub mod io;
-pub mod reconciler;
