+++ src/shim_layer/mod.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
 use crate::kubernetes_api_objects::{api_method::*, common::*, dynamic::*, error::*, resource::*};
-use crate::reconciler::exec::*;
+use crate::reconciler::exec::{io::*, reconciler::*};
 use builtin::*;
 use builtin_macros::*;
 use core::fmt::Debug;
